# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DynamicsERP API specification — Solar Projects ERP
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      parameters:
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Account" }
    post:
      operationId: createAccount
      tags: [accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "404":
          description: Not found
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    patch:
      operationId: updateAccount
      tags: [accounts]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
    delete:
      operationId: deleteAccount
      tags: [accounts]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }

  # ===== Contacts =====
...
  /quotations/{id}:
    get:
      operationId: getQuotation
      tags: [quotations]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/QuotationDetail" } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    patch:
      operationId: updateQuotation
      tags: [quotations]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/QuotationInput" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Quotation" } } } }
    delete:
      operationId: deleteQuotation
      tags: [quotations]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "204": { description: Deleted }
...
  /approval-rules:
    get:
      operationId: listApprovalRules
      tags: [approvals]
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ApprovalRule" } } } } }
    post:
      operationId: createApprovalRule
      tags: [approvals]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ApprovalRuleInput" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/ApprovalRule" } } } }
  /approval-rules/{id}:
    patch:
      operationId: updateApprovalRule
      tags: [approvals]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ApprovalRuleInput" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ApprovalRule" } } } }
    delete:
      operationId: deleteApprovalRule
      tags: [approvals]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "204": { description: Deleted }
  /approval-requests:
    get:
      operationId: listApprovalRequests
      tags: [approvals]
      parameters:
        - { in: query, name: assignedToMe, schema: { type: boolean } }
        - { in: query, name: status, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ApprovalRequest" } } } } }
  /approval-requests/{id}/approve:
    post:
      operationId: approveApprovalRequest
      tags: [approvals]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody:
        required: false
        content: { application/json: { schema: { $ref: "#/components/schemas/ApprovalActionInput" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ApprovalRequest" } } } }
  /approval-requests/{id}/reject:
    post:
      operationId: rejectApprovalRequest
      tags: [approvals]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody:
        required: false
        content: { application/json: { schema: { $ref: "#/components/schemas/ApprovalActionInput" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ApprovalRequest" } } } }

  # ===== Sales Orders =====
...
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchase-orders]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/PurchaseOrderDetail" } } } }
    patch:
      operationId: updatePurchaseOrder
      tags: [purchase-orders]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/UpdatePurchaseOrder" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/PurchaseOrderDetail" } } } }
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Account:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        industry: { type: ["string", "null"] }
        gstin: { type: ["string", "null"] }
        billingAddress: { type: ["string", "null"] }
        billingCity: { type: ["string", "null"] }
        billingState: { type: ["string", "null"] }
        billingPincode: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        accountType: { type: string, enum: [residential, commercial, industrial, government] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, name, accountType, createdAt]
    AccountInput:
      type: object
      properties:
        name: { type: string }
        industry: { type: ["string", "null"] }
        gstin: { type: ["string", "null"] }
        billingAddress: { type: ["string", "null"] }
        billingCity: { type: ["string", "null"] }
        billingState: { type: ["string", "null"] }
        billingPincode: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        accountType: { type: string, enum: [residential, commercial, industrial, government] }
        notes: { type: ["string", "null"] }
      required: [name, accountType]
...
    QuotationStatus:
      type: string
      enum: [draft, pending_approval, approved, rejected, sent, won, lost, cancelled]
    Quotation:
      type: object
      properties:
        id: { type: integer }
        quotationNumber: { type: string }
        leadId: { type: ["integer", "null"] }
        leadTitle: { type: ["string", "null"] }
        accountId: { type: ["integer", "null"] }
        accountName: { type: ["string", "null"] }
        contactId: { type: ["integer", "null"] }
        title: { type: string }
        status: { $ref: "#/components/schemas/QuotationStatus" }
        validUntil: { type: ["string", "null"], format: date }
        subtotal: { type: number }
        discountAmount: { type: number }
        taxableAmount: { type: number }
        gstAmount: { type: number }
        total: { type: number }
        notes: { type: ["string", "null"] }
        termsAndConditions: { type: ["string", "null"] }
        createdById: { type: ["integer", "null"] }
        createdByName: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, quotationNumber, title, status, subtotal, discountAmount, taxableAmount, gstAmount, total, createdAt, updatedAt]
    QuotationInput:
      type: object
      properties:
        leadId: { type: ["integer", "null"] }
        accountId: { type: ["integer", "null"] }
        contactId: { type: ["integer", "null"] }
        title: { type: string }
        validUntil: { type: ["string", "null"], format: date }
        discountAmount: { type: number, default: 0 }
        notes: { type: ["string", "null"] }
        termsAndConditions: { type: ["string", "null"] }
      required: [title]
    QuotationLineItem:
      type: object
      properties:
        id: { type: integer }
        quotationId: { type: integer }
        productId: { type: ["integer", "null"] }
        productName: { type: string }
        description: { type: ["string", "null"] }
        quantity: { type: number }
        unit: { type: string }
        unitPrice: { type: number }
        discountPct: { type: number }
        gstRate: { type: number }
        lineTotal: { type: number }
        position: { type: integer }
      required: [id, quotationId, productName, quantity, unit, unitPrice, discountPct, gstRate, lineTotal, position]
...
    QuotationDetail:
      allOf:
        - $ref: "#/components/schemas/Quotation"
        - type: object
          properties:
            lineItems:
              type: array
              items: { $ref: "#/components/schemas/QuotationLineItem" }
            approvalRequests:
              type: array
              items: { $ref: "#/components/schemas/ApprovalRequest" }
          required: [lineItems, approvalRequests]
...
    ApprovalRule:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        entityType: { type: string, enum: [quotation] }
        minAmount: { type: number }
        maxAmount: { type: ["number", "null"] }
        approverRole: { type: string }
        level: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, entityType, minAmount, approverRole, level, isActive, createdAt]
    ApprovalRuleInput:
      type: object
      properties:
        name: { type: string }
        entityType: { type: string }
        minAmount: { type: number }
        maxAmount: { type: ["number", "null"] }
        approverRole: { type: string }
        level: { type: integer }
        isActive: { type: boolean, default: true }
      required: [name, entityType, minAmount, approverRole, level]
    ApprovalRequestStatus:
      type: string
      enum: [pending, approved, rejected, cancelled]
    ApprovalRequest:
      type: object
      properties:
        id: { type: integer }
        entityType: { type: string }
        entityId: { type: integer }
        entityTitle: { type: ["string", "null"] }
        amount: { type: number }
        ruleId: { type: ["integer", "null"] }
        ruleName: { type: ["string", "null"] }
        approverRole: { type: string }
        approverId: { type: ["integer", "null"] }
        approverName: { type: ["string", "null"] }
        level: { type: integer }
        status: { $ref: "#/components/schemas/ApprovalRequestStatus" }
        comments: { type: ["string", "null"] }
        actionedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
      required: [id, entityType, entityId, amount, approverRole, level, status, createdAt]
    ApprovalActionInput:
      type: object
      properties:
        comments: { type: ["string", "null"] }
...
    PurchaseOrder:
      type: object
      properties:
        id: { type: integer }
        poNumber: { type: string }
        vendorId: { type: integer }
        vendorName: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
        projectName: { type: ["string", "null"] }
        title: { type: string }
        status: { type: string }
        orderDate: { type: string, format: date }
        expectedDeliveryDate: { type: ["string", "null"], format: date }
        supplyType: { type: string }
        placeOfSupply: { type: string }
        subtotal: { type: number }
        discountAmount: { type: number }
        taxableAmount: { type: number }
        cgstAmount: { type: number }
        sgstAmount: { type: number }
        igstAmount: { type: number }
        total: { type: number }
        notes: { type: ["string", "null"] }
        termsAndConditions: { type: ["string", "null"] }
        sentAt: { type: ["string", "null"], format: date-time }
        approvedById: { type: ["integer", "null"] }
        approvedByName: { type: ["string", "null"] }
        approvedAt: { type: ["string", "null"], format: date-time }
        createdById: { type: ["integer", "null"] }
        createdByName: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, poNumber, vendorId, title, status, orderDate, supplyType, placeOfSupply, subtotal, discountAmount, taxableAmount, cgstAmount, sgstAmount, igstAmount, total, createdAt, updatedAt]
    PoLineItem:
      type: object
      properties:
        id: { type: integer }
        purchaseOrderId: { type: integer }
        productName: { type: string }
        description: { type: ["string", "null"] }
        hsnCode: { type: ["string", "null"] }
        quantity: { type: number }
        receivedQuantity: { type: number }
        unit: { type: string }
        unitPrice: { type: number }
        discountPct: { type: number }
        gstRate: { type: number }
        lineTotal: { type: number }
      required: [id, purchaseOrderId, productName, quantity, receivedQuantity, unit, unitPrice, discountPct, gstRate, lineTotal]
    PurchaseOrderDetail:
      allOf:
        - $ref: "#/components/schemas/PurchaseOrder"
        - type: object
          properties:
            lineItems: { type: array, items: { $ref: "#/components/schemas/PoLineItem" } }
          required: [lineItems]
...
    UpdatePurchaseOrder:
      type: object
      properties:
        title: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
        expectedDeliveryDate: { type: ["string", "null"], format: date }
        placeOfSupply: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        termsAndConditions: { type: ["string", "null"] }
...